Refactored SDate and SDuration into their own files
