+++ packages/squiggle-lang/src/utility/SDate.ts
@@ -1,114 +1,7 @@
 import { REOther } from "../errors/messages.js";
+import { SDuration, durationUnits } from "./SDuration.js";
 import { Ok, result, Err } from "./result.js";
 import * as Result from "./result.js";
-
-// Stores in Unix milliseconds
-export type Duration = number;
-
-// TODO - should we just use date-fns instead? Dates are hard to implement correctly.
-
-const durationUnits = {
-  Second: 1000,
-  Minute: 60 * 1000,
-  Hour: 60 * 60 * 1000,
-  Day: 24 * 60 * 60 * 1000,
-  Year: 24 * 60 * 60 * 1000 * 365.25,
-} as const;
-
-export class SDuration {
-  private static durationUnits = {
-    Second: 1000,
-    Minute: 60 * 1000,
-    Hour: 60 * 60 * 1000,
-    Day: 24 * 60 * 60 * 1000,
-    Year: 24 * 60 * 60 * 1000 * 365.25,
-  } as const;
-
-  constructor(private ms: number) {
-    this.ms = ms;
-  }
-
-  static fromMs(f: number): SDuration {
-    return new SDuration(f);
-  }
-
-  static fromMinutes(minutes: number): SDuration {
-    return new SDuration(minutes * SDuration.durationUnits.Minute);
-  }
-
-  static fromHours(hours: number): SDuration {
-    return new SDuration(hours * SDuration.durationUnits.Hour);
-  }
-
-  static fromDays(days: number): SDuration {
-    return new SDuration(days * SDuration.durationUnits.Day);
-  }
-
-  static fromYears(years: number): SDuration {
-    return new SDuration(years * SDuration.durationUnits.Year);
-  }
-
-  toMs(): number {
-    return this.ms;
-  }
-
-  toMinutes(): number {
-    return this.ms / SDuration.durationUnits.Minute;
-  }
-
-  toHours(): number {
-    return this.ms / SDuration.durationUnits.Hour;
-  }
-
-  toDays(): number {
-    return this.ms / SDuration.durationUnits.Day;
-  }
-
-  toYears(): number {
-    return this.ms / SDuration.durationUnits.Year;
-  }
-
-  toString(): string {
-    const units: [number, string][] = [
-      [SDuration.durationUnits.Year, "year"],
-      [SDuration.durationUnits.Day, "day"],
-      [SDuration.durationUnits.Hour, "hour"],
-      [SDuration.durationUnits.Minute, "minute"],
-      [SDuration.durationUnits.Second, "second"],
-    ];
-
-    for (const [unitValue, unitName] of units) {
-      if (Math.abs(this.ms) >= unitValue) {
-        const value = this.ms / unitValue;
-        const suffix = value !== 1.0 ? "s" : "";
-        return `${value.toPrecision(3)} ${unitName}${suffix}`;
-      }
-    }
-
-    return `${this.ms.toFixed()} ms`;
-  }
-
-  add(other: SDuration): SDuration {
-    return new SDuration(this.ms + other.ms);
-  }
-
-  subtract(other: SDuration): SDuration {
-    return new SDuration(this.ms - other.ms);
-  }
-
-  divideBySDuration(divisor: SDuration): number {
-    return this.ms / divisor.toMs();
-  }
-
-  divideByNumber(divisor: number): SDuration {
-    return new SDuration(this.ms / divisor);
-  }
-
-  //Assumes the passed-in number is unitless
-  multiply(num: number): SDuration {
-    return SDuration.fromMs(this.toMs() * num);
-  }
-}
 
 export class SDate {
   constructor(public value: Date) {
