+++ packages/squiggle-lang/src/value/index.ts
@@ -6,7 +6,8 @@ import {
   REOther,
 } from "../errors/messages.js";
 import { Lambda } from "../reducer/lambda.js";
-import { SDate, SDuration } from "../utility/DateTime.js";
+import { SDate } from "../utility/SDate.js";
+import { SDuration } from "../utility/SDuration.js";
 import { ImmutableMap } from "../utility/immutableMap.js";
 import { DateRangeDomain, Domain, NumericRangeDomain } from "./domain.js";
 import { shuffle } from "../utility/E_A.js";
